prism: A2.GR — gracefulShutdown helper for bwrap/sandbox-exec
Issue #1149 part 3 of 4. Extracts the SIGTERM-then-grace-then-SIGKILL body
from bwrapIsolator.Shutdown into a free function gracefulShutdown(*exec.Cmd,
gracePeriod) used by both bwrap and sandbox-exec Shutdown methods (A2 §3.7).

bwrapIsolator.Shutdown is reduced to a thin call into gracefulShutdown.
sandboxExecIsolator gains a cmd field and a Run implementation that stores
the *exec.Cmd, so its Shutdown can deliver the same SIGTERM/SIGKILL sequence
when invoked. In production both Run paths remain unused (the bwrap and
sandbox-exec children are owned by cmd/agent_run.go and
cmd/agent_run_sandbox_exec_darwin.go respectively); the helper is reachable
via the Isolator.Shutdown() public API and gives both modes a real graceful
shutdown body for tests and future Manager.Run paths.

Podman's Shutdown stays separate (must-differ per A2 §3.7) — it drives
`podman stop --time 10` + `podman rm --force` rather than direct signal
delivery.
